improved cookie management
