+++ frontend/src/App.vue
@@ -1,11 +1,12 @@
 <script setup lang="ts">
 import Notification from "@/components/common/Notification.vue";
 import api from "@/services/api/index";
+import userApi from "@/services/api/user";
+import storeAuth from "@/stores/auth";
 import storeConfig from "@/stores/config";
 import storeHeartbeat from "@/stores/heartbeat";
 import storeNotifications from "@/stores/notifications";
-import userApi from "@/services/api/user";
-import storeAuth from "@/stores/auth";
+import Cookies from "js-cookie";
 import { storeToRefs } from "pinia";
 import { onBeforeMount } from "vue";
 import router from "./plugins/router";
@@ -18,25 +19,36 @@ const auth = storeAuth();
 const configStore = storeConfig();
 
 onBeforeMount(async () => {
-  await api.get("/heartbeat").then(({ data: data }) => {
-    heartbeat.set(data);
-    if (heartbeat.value.SETUP_WIZARD) {
-      router.push({ name: "setup" });
-    }
-  });
+  await api
+    .get("/heartbeat")
+    .then(async ({ data: data }) => {
+      heartbeat.set(data);
+      if (heartbeat.value.SETUP_WIZARD) {
+        router.push({ name: "setup" });
+      } else {
+        await userApi
+          .fetchCurrentUser()
+          .then(({ data: user }) => {
+            auth.setUser(user);
+          })
+          .catch((error) => {
+            console.error(error);
+          });
 
-  await userApi
-    .fetchCurrentUser()
-    .then(({ data: user }) => {
-      auth.setUser(user);
+        await api.get("/config").then(({ data: data }) => {
+          configStore.set(data);
+        });
+      }
     })
-    .catch((error) => {
-      console.error(error);
+    .catch(() => {
+      const allCookies = Cookies.get(); // Get all cookies
+      for (let cookie in allCookies) {
+        Cookies.remove(cookie); // Remove each cookie
+      }
+      router.push({
+        name: "login",
+      });
     });
-
-  await api.get("/config").then(({ data: data }) => {
-    configStore.set(data);
-  });
 });
 </script>
