+++ frontend/src/components/common/Navigation/SettingsDrawer.vue
@@ -14,6 +14,7 @@ import { useDisplay } from "vuetify";
 const navigationStore = storeNavigation();
 const router = useRouter();
 const auth = storeAuth();
+const { user, scopes } = storeToRefs(auth);
 const emitter = inject<Emitter<Events>>("emitter");
 const { activeSettingsDrawer } = storeToRefs(navigationStore);
 const { smAndDown } = useDisplay();
@@ -43,8 +44,8 @@ async function logout() {
       <v-list-img>
         <v-img
           :src="
-            auth.user?.avatar_path
-              ? `/assets/romm/assets/${auth.user?.avatar_path}?ts=${auth.user?.updated_at}`
+            user?.avatar_path
+              ? `/assets/romm/assets/${user?.avatar_path}?ts=${user?.updated_at}`
               : defaultAvatarPath
           "
           cover
@@ -52,8 +53,8 @@ async function logout() {
         </v-img>
       </v-list-img>
       <v-list-item
-        :title="auth.user?.username"
-        :subtitle="auth.user?.role"
+        :title="user?.username"
+        :subtitle="user?.role"
         class="mb-1 text-shadow text-white"
       >
       </v-list-item>
@@ -63,13 +64,13 @@ async function logout() {
         >UI Settings</v-list-item
       >
       <v-list-item
-        v-if="auth.scopes.includes('platforms.write')"
+        v-if="scopes.includes('platforms.write')"
         append-icon="mdi-table-cog"
         :to="{ name: 'management' }"
         >Library Management
       </v-list-item>
       <v-list-item
-        v-if="auth.scopes.includes('users.write')"
+        v-if="scopes.includes('users.write')"
         :to="{ name: 'administration' }"
         append-icon="mdi-security"
         >Administration</v-list-item
